rearrange code a bit
