+++ UI/Dialogs/Editors/propeditor/StreamEditor.h
